made OCD happy this really is quite an unnecessary commit, but I felt it was necessary to do it.
there was no point to this commit, therefore it was pointless.
it was so pointless, that nobody cared about it, except it screwed with someone's merge, so they had to mess around with it.

actually, the point was for me to test this commit message system.
perhaps it will work.

you can never be sure