changed a bunch of things
> album rendering is prettier now
can be changed, of course, and I think the background could be slightly transparent,
but I like it.
> photo caption rendering in the album renderer is slightly better
the caption no longer fills the entire bottom of the div, just the max-width, which
is 40px now (larger numbers may need some attention to make sure that the ID works
correctly.
> some other tweaks with some other things, but album rendering is the biggest
thing done.