Modify Gemfile, fix up controller some more, remove silly self calls
self is not necessary to be called in a model, because it is just redundant typing.

I added `gem "minitest"` to the Gemfile because ActiveSupport complained about not finding it.  It is a really weird problem that does not really make sense.  I did have it installed; I could sucessfully call `require 'minitest/unit'` in irb.  Oh well, the extra line won't harm anyone, right?  I used http://stackoverflow.com/questions/12501067/cant-make-the-rails-console-run-on-cygwin to find the answer.