Fix events iCalendar file (#63)
The gem we use for generating the iCal file does not handle
ActiveRecord's DateTime class properly (it has its own custom one, I
imagine). Apparently, you can pass it a pre-formatted string, which is
what I did to bypass this issue. If there is a better way to do this, I
am open to using it, but this works for now.